refactoring and disabled logging by default
