+++ Log.py
@@ -1,14 +1,16 @@
 import os
 
 USERNAME = os.getlogin()
-
+from Settings import Settings
 
 with open(f'C:/Users/{USERNAME}/Documents/Ableton/User Library/Remote Scripts/log.txt', 'a') as f:
-    f.write('====================\n')
+    if Settings.LOGGING:
+        f.write('====================\n')
 log_num = 0
 
 def log(message):
     global log_num
     with open(f'C:/Users/{USERNAME}/Documents/Ableton/User Library/Remote Scripts/log.txt', 'a') as f:
-        f.write(str(log_num) + ' ' + message + '\n')
+        if Settings.LOGGING:
+            f.write(str(log_num) + ' ' + message + '\n')
     log_num += 1
