+++ PreciseButtonSliderElement.py
@@ -1,6 +1,6 @@
 
 from _Framework.ButtonSliderElement import ButtonSliderElement
-from .Test import log
+from .Log import log
 SLIDER_MODE_SINGLE = 0
 SLIDER_MODE_VOLUME = 1
 SLIDER_MODE_PAN = 2
